+++ lib/api_thread.rb
@@ -1,5 +1,6 @@
 require 'restful_model'
 require 'time_attr_accessor'
+require 'mixins'
 
 module Inbox
   class Thread < RestfulModel
@@ -8,7 +9,6 @@ module Inbox
     parameter :subject
     parameter :participants
     parameter :snippet
-    parameter :tags
     parameter :message_ids
     parameter :draft_ids
     parameter :labels
@@ -18,6 +18,8 @@ module Inbox
     parameter :has_attachments
     time_attr_accessor :last_message_timestamp
     time_attr_accessor :first_message_timestamp
+
+    include ReadUnreadMethods
 
     def inflate(json)
       super
@@ -45,37 +47,6 @@ module Inbox
 
     def drafts
       @drafts ||= RestfulModelCollection.new(Draft, @_api, {:thread_id=>@id})
-    end
-
-    def update_tags!(tags_to_add = [], tags_to_remove = [])
-      update('PUT', '', {
-        :add_tags => tags_to_add,
-        :remove_tags => tags_to_remove
-      })
-    end
-
-    def mark_as_read!
-      update_tags!([], ['unread'])
-    end
-
-    def mark_as_seen!
-      update_tags!([], ['unseen'])
-    end
-
-    def archive!
-      update_tags!(['archive'], ['inbox'])
-    end
-
-    def unarchive!
-      update_tags!(['inbox'], ['archive'])
-    end
-
-    def star!
-      update_tags!(['starred'], [''])
-    end
-
-    def unstar!
-      update_tags!([], ['starred'])
     end
 
     def as_json(options = {})