+++ lib/tag.rb
@@ -1,9 +1,0 @@
-require 'restful_model'
-
-module Inbox
-  class Tag < RestfulModel
-
-    parameter :name
-
-  end
-end